refactor(realtime): move socket to dedicated realtime dir
