+++ backend/server.js
@@ -1,7 +1,7 @@
 import dotenv from "dotenv";
 import { connectDB } from "./src/db.js";
 import app from "./src/app.js";
-import { setupSocket } from "./src/utils/socket.js";
+import { setupSocket } from "./src/realtime/io.js";
 
 dotenv.config();
